# NI-394 - reduce redraw