<!-- markdownlint-disable MD041 -->

### Background

our swiftUI views are being redrawn too many times resulting in really slow loading times for more complex views.
here's a couple things i've identified that works and has cut down rendering time by half (rendering the bottomsheet shown as benchmark)
- reducing the use of constants when initializing bindings, can't find any literature on it, but this can trigger a redraw when we initialize a view that has a binding with a constant.
- @Published property triggers the view to redraw indiscriminately even if the value is same, introducing a new property wrapper @LazyPublished that only triggers redraw when the wrapped value is different (thanks to [swift evolution 0258](https://github.com/swiftlang/swift-evolution/blob/main/proposals/0258-property-wrappers.md#referencing-the-enclosing-self-in-a-wrapper-type)).

| | Before | After |
|:------:|:------:|:------:|
| time to display | 500ms  | 200ms |

[bottomsheet layout gist](https://gist.github.com/WeiYewT/02139f9a05161916745265413e4949c9)

further investigation needed if we can decouple BasicTextComponent with the progress state indicator, as that has alot of state management that is not needed most of the time.

Fixes [([issue](https://rokt.atlassian.net/browse/NI-394))]

### What Has Changed

List out what has changed as a result of this PR.

### How Has This Been Tested?

Describe the tests that you ran to verify your changes.

### Notes

Add any notes or extra information here that might be useful to the reviewer if applicable i.e. links to documentation/blogs or dashboards.

### Checklist

- [x] I have performed a self-review of my own code.
- [x] I have made corresponding changes to the documentation.
- [ ] I have added tests that prove my fix is effective or that my feature works.
- [x] New and existing unit tests pass locally with my changes.